API Reference

GACODE

GACODE.FluxSolution — Type

ENERGY_FLUX_e::T
ENERGY_FLUX_i::T
PARTICLE_FLUX_e::T
PARTICLE_FLUX_i::Vector{T}
STRESS_TOR_i::T

Structure used to store fluxes information from GA code in Gyrobohm units

GACODE.InputGACODE — Type

InputGACODE(dd::IMAS.dd)

Creates an input_gacode from dd

GACODE.bunit — Function

bunit(eqt1d::IMAS.equilibrium__time_slice___profiles_1d)

Calculate bunit from equilibrium

GACODE.c_s — Function

c_s(cp1d::IMAS.core_profiles__profiles_1d)

Sounds speed in [cm/s]

GACODE.flux_gacode_to_imas — Function

flux_gacode_to_imas(
    flux_types::Tuple{Vararg{Symbol}},
    flux_solutions::Vector{FluxSolution{T}},
    m1d::IMAS.core_transport__model___profiles_1d,
    eqt::IMAS.equilibrium__time_slice,
    cp1d::IMAS.core_profiles__profiles_1d
)

Normalizes specified transport fluxes output by GA code via gyrobohm normalization and Miller volume correction

GACODE.gyrobohm_energy_flux — Function

gyrobohm_energy_flux(cp1d::IMAS.core_profiles__profiles_1d, eqt::IMAS.equilibrium__time_slice)

Gyrobohm energy flux

GACODE.gyrobohm_momentum_flux — Function

gyrobohm_momentum_flux(cp1d::IMAS.core_profiles__profiles_1d, eqt::IMAS.equilibrium__time_slice)

Gyrobohm momentum flux

GACODE.gyrobohm_particle_flux — Function

gyrobohm_particle_flux(cp1d::IMAS.core_profiles__profiles_1d, eqt::IMAS.equilibrium__time_slice)

Gyrobohm particle flux

Missing docstring.

Missing docstring for load. Check Documenter's build log for details.

GACODE.pick_ion_flux — Function

pick_ion_flux(ion_fluxes::AbstractVector{T}, kk::Int) where {T<:Real}

Select which ion flux to take

GACODE.r_min_core_profiles — Function

r_min_core_profiles(eqt::IMAS.equilibrium__time_slice, rho_tor_norm::AbstractVector)

Geometric minor radius in [cm] evaluated

GACODE.rho_s — Function

rho_s(cp1d::IMAS.core_profiles__profiles_1d, eqt::IMAS.equilibrium__time_slice)

sound gyro radius in [cm]

Missing docstring.

Missing docstring for save. Check Documenter's build log for details.

GACODE.volume_prime_miller_correction — Function

volume_prime_miller_correction(eqt::IMAS.equilibrium__time_slice)

Correction to account for transformation from Miller r grid in GA code equilibrium to Psi grid in FUSE equilibrium